# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is official backend documentation for the fngc software",
        "title": "FNGC Backend Service",
        "contact": {
            "name": "Endy Apinageri",
            "email": "apinaendy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "petstore.swagger.io",
    "basePath": "/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Accept JSON data of User Login objects and returns valid response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Handle unique Unique User Login",
                "parameters": [
                    {
                        "description": "The Tutor Login Data",
                        "name": "StudentData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.LoginData"
                        }
                    },
                    "400": {
                        "description": "Check Response Message",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseBody"
                        }
                    }
                }
            }
        },
        "/auth/resetpassword": {
            "post": {
                "description": "Accept JSON data of User Reset password objects and returns valid response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Handle resetting a user password",
                "parameters": [
                    {
                        "description": "The User Data",
                        "name": "StudentData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResetPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.ResetPassword"
                        }
                    },
                    "400": {
                        "description": "Check Response Message",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseBody"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Accept JSON data of Student User objects and returns valid response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Handle unique User Registration",
                "parameters": [
                    {
                        "description": "The Student Registration Data",
                        "name": "StudentData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserRegistrationData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.UserRegistrationData"
                        }
                    },
                    "400": {
                        "description": "Check Response Message",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseBody"
                        }
                    }
                }
            }
        },
        "/auth/verifyotp": {
            "post": {
                "description": "Accept JSON data of User Reset password objects and returns valid response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Handle verifying user otp",
                "parameters": [
                    {
                        "description": "The User Verification Data",
                        "name": "StudentData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VerifyUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.VerifyUser"
                        }
                    },
                    "400": {
                        "description": "Check Response Message",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseBody"
                        }
                    }
                }
            }
        },
...
        "/user/studyabroad/": {
            "post": {
                "description": "Accept JSON data for study details objects and returns valid response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hybrid APIs"
                ],
                "summary": "Handle data to process abroad study",
                "parameters": [
                    {
                        "description": "The Task Data",
                        "name": "StudyAbroadData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StudyAbroad"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.StudyAbroad"
                        }
                    },
                    "400": {
                        "description": "Check Response Message",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseBody"
                        }
                    }
                }
            }
        },
        "/user/task/": {
            "post": {
                "description": "Accept JSON data for examination objects and returns valid response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hybrid APIs"
                ],
                "summary": "Handle adding task for users",
                "parameters": [
                    {
                        "description": "The Task Data",
                        "name": "TaskData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    },
                    "400": {
                        "description": "Check Response Message",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.LoginData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "models.ResetPassword": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                },
                "reset_password_link": {
                    "type": "string"
                }
            }
        },
        "models.ResponseBody": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object"
                },
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.StudyAbroad": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "country_of_study": {
                    "type": "string"
                },
                "course": {
                    "type": "string"
                },
                "degree": {
                    "type": "string"
                },
                "degree_level": {
                    "type": "string"
                },
                "degree_prgramme": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "educational_level": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enrollement_year": {
                    "type": "string"
                },
                "enrollment_term": {
                    "type": "string"
                },
                "financial_status": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "graduation_year": {
                    "type": "string"
                },
                "high_school": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "marital_status": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "study_model": {
                    "type": "string"
                },
                "university_name": {
                    "type": "string"
                },
                "year_of_admission": {
                    "type": "string"
                }
            }
        },
        "models.Task": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "task": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.UserRegistrationData": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
        "models.VerifyUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "verification_otp": {
                    "type": "string"
                }
            }
        }
    }
}